+++ test/conformance/api/resources_test.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package conformance
+package api
 
 import (
 	"fmt"
@@ -33,7 +33,7 @@ import (
 
 func TestCustomResourcesLimits(t *testing.T) {
 	t.Parallel()
-	clients := setup(t)
+	clients := test.Setup(t)
 
 	t.Log("Creating a new Route and Configuration")
 	resources := corev1.ResourceRequirements{
